+++ Kernel/include/queueADT.h
@@ -1,7 +1,7 @@
 #ifndef _QUEUE_ADT_H_
 #define _QUEUE_ADT_H_
 
-#include <stddef.h>
+#include <stdint.h>
 
 typedef struct queueNode {
     void *data;
@@ -11,7 +11,7 @@ typedef struct queueNode {
 typedef struct queueCDT {
     queueNode *head;
     queueNode *tail;
-    size_t size;
+    uint64_t size;
 } *queueADT;
 
 queueADT createQueue(void);
@@ -22,7 +22,7 @@ void *dequeue(queueADT q);
 void *dequeueValue(queueADT q, void *elem);
 
 int queueIsEmpty(queueADT q);
-size_t queueSize(queueADT q);
-size_t countOccurrences(queueADT q, void *elem);
+uint64_t queueSize(queueADT q);
+uint64_t countOccurrences(queueADT q, void *elem);
 
 #endif
